example_project/requirements.txt
-r ../requirements.txt

django-indexer==0.3.0
django-lazysignup==0.7.0
django-paging==0.2.4
django-sentry==1.8.5
django-templatetag-sugar==0.1
gunicorn==0.12.2

# Django 1.3 support plox
-e git+https://github.com/pydanny/django-uni-form.git@32baac656ce3116aa83d60e32eb536aca6483dd6#egg=django-uni-form
